Squashed commit of the following:
commit bd69ec24327a4a2260abb90a3dea50038394c054
Author: bhanu.k <bhanu.kamapantula@gramener.com>
Date:   Thu Feb 11 20:31:06 2021 +0530

    ENH: form rendering using JSON

    - deleted images in introduction directory, artifact from an older version of the app
    - introduces duplicate/delete actions along with a field move option
    - JSON form rendering used in form creation (`/create`) and form rendering (`/form/ID`, `/view/ID`)

commit 240e07155a117499abb8e9e78631c30fef99494b
Author: bhanu.k <bhanu.kamapantula@gramener.com>
Date:   Thu Feb 11 11:23:35 2021 +0530

    ENH: redirect on form creation with form id

commit a51d1fe845363d551bad4131e85ffcb31f3ce620
Author: bhanu.k <bhanu.kamapantula@gramener.com>
Date:   Thu Feb 11 11:20:06 2021 +0530

    FXD: delete form_ID.db only if it exists

+++ gramex/apps/uifactory/form_builder.py
@@ -46,7 +46,8 @@ def after_publish(handler, data):
     elif handler.request.method == 'DELETE':
         _id = handler.get_argument('id')
         gramex.data.delete(url=var.FORMS_URL, table=var.FORMS_TABLE, args=handler.args, id=['id'])
-        os.remove(os.path.join(var.GRAMEXDATA, 'uifactory', f'form_{_id}.db'))
+        if(os.path.exists(os.path.join(var.GRAMEXDATA, 'uifactory', f'form_{_id}.db'))):
+            os.remove(os.path.join(var.GRAMEXDATA, 'uifactory', f'form_{_id}.db'))
 
 
 @handler
